Completed the view details methods

+++ m2-api-capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/TransferDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface TransferDAO {
 
-    public List<Transfer> retrieveTransferHistory(int userId);
+    public List<Transfer> retrieveTransferHistory(Long userId);
     public List<Transfer> retrievePendingRequests(int userId);
     public Transfer sendBucks (Transfer transfer);
 }
